+++ restaurant-miniprogram/miniprogram/config/index.js
@@ -1,5 +1,5 @@
 //var API_BASE_URL = 'http://localhost:8080'
-var API_BASE_URL = 'http://192.168.0.101:8080'
+var API_BASE_URL = 'http://192.168.0.105:8080'
 
 var STORAGE_KEYS = {
   token: 'token',